requirements.txt
# pip install torch==2.0.1+cu117 torchvision==0.15.2+cu117 torchaudio==2.0.2+cu117 --extra-index-url https://download.pytorch.org/whl/cu117

numpy
matplotlib
Pillow
pandas
seaborn
scikit-learn
tqdm
pudb==2019.2
pyspng==0.1.0 
blobfile
pyproject-toml
mpi4py
gradio
albumentations==1.3.0
opencv-python==4.8.0.74
opencv-contrib-python==4.3.0.36
opencv-python-headless==4.5.5.64
imageio==2.9.0
imageio-ffmpeg==0.4.2
pytorch-lightning
omegaconf
test-tube>=0.7.5
streamlit==1.12.1
einops==0.6.1
transformers==4.25.1
webdataset==0.2.5
kornia==0.6
open_clip_torch==2.20.0
invisible-watermark>=0.1.5
streamlit-drawable-canvas==0.8.0
torch-fidelity==0.3.0
torchmetrics==0.6.0
torchinfo
timm==0.4.12
addict==2.4.0
yapf==0.32.0
prettytable==3.6.0
safetensors
basicsr==1.4.2
taming-transformers-rom1504
hydra
hydra-optuna-sweeper
mlflow
accelerate
tokenizers
sentencepiece~=0.1.97
ftfy~=6.1.1
beautifulsoup4~=4.11.1
git+https://github.com/openai/CLIP.git
controlnet_aux
salesforce-lavis==1.0.2
lpips